// src/app/models/hitter.ts
export interface Hitter {
    ab: number;
    actual_level: string;
    age: number;
    avg: number;
    babip: number;
    bb: number;
    bb_rate: number;
    contact_rate: number;
    cs: number;
    d: number;
    fb_rate: number;
    g: number;
    gb_fb: number;
    gb_rate: number;
    h: number;
    hard_hit_rate: number;
    hr: number;
    hr_fb_rate: number;
    iso: number;
    ld_rate: number;
    level: string;
    name: string;
    obp: number;
    ops: number;
    pa: number;
    player_id: number;
    position: string;
    r: number;
    rbi: number;
    roster_status: string;
    sb: number;
    sb_rate: number;
    slg: number;
    so: number;
    so_rate: number;
    speed: number;
    sw_strike_rate: number;
    t: number;
    woba: number;
    wraa: number;
    wrc_plus: number;
    xwoba: number;
}