// src/app/models/pitcher.ts
export interface Pitcher {
    ab: number;
    actual_level: string;
    age: number;
    babip: number;
    barrel_rate: number;
    bb: number;
    bb_9: number;
    bb_rate: number;
    bf: number;
    contact_rate: number;
    er: number;
    era: number;
    fb_rate: number;
    fip: number;
    g: number;
    gb_fb_rate: number;
    gb_rate: number;
    gs: number;
    h: number;
    hard_hit_rate: number;
    hr: number;
    hr_fb_rate: number;
    ip: number;
    k_9: number;
    k_bb: number;
    l: number;
    ld_rate: number;
    level: string;
    name: string;
    player_id: number;
    position: string;
    r: number;
    rar: number;
    roster_status: string;
    s: number;
    so: number;
    so_rate: string;
    sw_strike_rate: number;
    w: number;
    war: number;
    whip: number;
    xfip: number;
}